Built and tested path_maker
Seems to work. But since I don't know how to run tests from a separate
folder yet I still have to resort to idiotic test protocols.
So, I had to copy the path_maker.py in to the test folder to check
how it performs. 'waste.py' is to store code not to be used in the code
in the forseeable future. 'path_maker.txt' has changed due to the
experiments i had to perform while formulating code for its namesake
python file.